realex_offsite: update avs code generation
The BILLING_CODE field is currently only formatted correctly for GB AVS
checks. This is where only the digits from the post code and the first
line of the address are submitted with a “|” between. For example:

<input type="hidden" name="BILLING_CODE" value="R90|123">

However, for US and CA the full post code and address should be submitted.
For example:

<input type="hidden" name="BILLING_CODE" value="50001|Flat 123">

We just need to introduce code here to only remove the non-digits when
the country is GB. For all other countries the full zip and first line
of the address should be submitted.