[Bitpay] Fixes an issue where notification couldn't be acknowledged
Previous changes to BitPay https://github.com/activemerchant/offsite_payments/pull/328 introduced a bug since the payload from an API and the payload present on the api aren't similar. This will set the params of the notification to `JSON.parse(body)['data']` if there is a `data` key or will set it to `JSON.parse(body) otherwise.