Lock Gemfile-shopify to money 0.14.x
The `Gemfile.shopify` file doesn't specify a version of money to use.
I'm not entirely sure _why_ that's the case, but it seems like an
unnecessary risk to me that can just lead to headaches.

A particular headache I'm currently experiencing (and that's
prompting me to lock this) is that Money 0.15 was just released
and several tests began to fail because of that.

Let's keep the minor version locked for now to prevent breakages
introduced by functionality changes.