Automatically eager load all integrations when in Rails context
In production mode it's preferable to load all the code upfront
even if it means loading some code that won't be used.

This is because loading the code on demand cause the first request
to use it to be slower, and to flush various VM caches such
as the global constant cache. It also degrade Copy on Write
performance for forking setups.